requirements.txt
pillow
numpy
bs4
python-dotenv
pymongo
selenium-wire
requests
selenium
packaging